Moved instruction matrix
